+++ d2/src/main/java/com/linkedin/d2/balancer/clients/DynamicClient.java
@@ -157,6 +157,7 @@ public class DynamicClient extends AbstractClient implements D2Client
 
       callback.onSuccess(None.none());
     });
+    TimingKey.unregisterKey(TIMING_KEY);
   }
 
   @Override